Websocket video working
Wrote a script to segment and fragment the sample video into ones that can be played by MSE. Added to the Readme the command to fragment. Added the proper codecs to the js so that it can read/play the videos. Made the buffer size larger as some of the video files were almost 400KB. Properly Segmented and fragmented files added for demo.